requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_bcrypt_1762497769/work
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1761203062999/work
click==8.3.1
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1770772381886/work
exceptiongroup==1.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.5
pydantic_core==2.41.5
python-bcrypt==0.3.2
python-dotenv==1.2.2
python-multipart==0.0.22
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
uvicorn==0.42.0
